added delete route for token cotoller and did other cleaning up
