+++ retina-api/retina-api/retina-api.csproj
@@ -107,7 +107,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App_Start\WebApiConfig.cs" />
-    <Compile Include="Controllers\AddController.cs" />
     <Compile Include="Controllers\BrandsController.cs" />
     <Compile Include="Controllers\OwnersController.cs" />
     <Compile Include="Controllers\StatusController.cs" />
